add event changes & shift verifyToken to auth middleware
